Update the dock item to for title button signals
This updates the dock item to proxy the signals from the internal title
bar. It allows the parent dock container to attach to the proxied
signals without worrying about the user changing the title bar widget at
a later time. It also delegates the responsibility for closing the item
to the parent container.