qt: fix issue #206
Previously when a widget was unparented from a QtContainer
subclass the layout_container member was not cleared. This
could lead to bugs when a relayout of the widget was
requested as it would try to relayout its old parent.
If the new one is not a Container the member is not updated.
Furthermore, the old  parent layout can be out of sync if
it is not displayed at the time, which is turns can lead to
an index error as a child appears missing.

I believe this fix is the cheapest one as it does not attempt
to change the layout of the parent.