core : fixes for closure in comprehensions
Wrap functions defined inside operators or declarative function to call
them with their scope of definition. This allows to handle properly
comprehensions and lambdas (whose body is now properly traced).
Also ensure that we compile the body of the :: operator as a function
to properly handle closure.

Bump the compiler version to 26