Several README.rst improvements (#271)
* Add some more bling to the ReadMe file

Would like to add a License and Python versions from PyPI, but not specified.

Also considered StackOverflow link.

* Clean up appearance of presentations

Removed large thumbnail in Welcome.
Added small thumbnails to both links in talks section.
Gave authors' full names.
Made it clearer they were links.

* Explicitly state versions supported.

This should also go in an appropriate file where PyPI can find it.

* Add link to license information

Would like to add this as a badge, but I think that has to wait until the next release to PyPI. PyPI hasn't yet seen the modification to setup.py which includes this info.

* Trivial spelling error in comment.

* Justify the examples

Add context to each of the examples justifying their existence.
Link to appropriate sections of documentation for more.
Push that the documentation is great at sample code.

* Merge Help and Learn More

Link to particular chapters that beginners will need.

* Substantial rewrite of the benefits of Enaml.

* Fix indenting

I don't know the secret for sublists without it thinking it is a quote.

* Fix flightless bird reference.

Also added links.

* Add Stackoverflow links.

* Change order of examples

So the narrative flow follows more easily.

Trivial extra changes.

* Formatting of next lists fixed.

RST is finicky compared to Markdown!

* Give the examples their own subsections

* Remove the StackOverflow shield

It is nice when it works, but is too unreliable.

The link is only loading an image about one time in three: https://img.shields.io/stackexchange/stackoverflow/t/enaml.svg

* Correct reference to PyPi in shield link

* Use action-shot for alerts example.

Replaced the screenshot of the dock items alerts example with one that shows the alerts in action - i.e. has coloured title bars, as this is the reason for the example. The previous screen shot showed no difference between the alert levels.

* Tightened up unecessary whitespace & underlines

No visual changes.