#282: Don't let pyqt5 5.11.2 be used. (#283)
pyqt5 5.11.2 has a (packaging?) bug and does not work on Windows 3.5. It is assumed that this will be fixed in the next dot release.
Simply avoid using it for all Python versions (even though it works fine on other versions).

Also, more/cleaner debug output and avoid hardcoding the APPVEYOR_BUILD_DIR; it actually changes!