Document the tools directory
Add new documentation to cover the usage of all the syntax-based tools in the tools directory.

Note: A lot of this is based on research, not actual usage. I think this is a lot better than nothing, though.

Also, it would be nice to somehow incorporate the two README.rst files into the documentation, but I have left that for the future.