Allow explicit setting of the application name
The argument to QApplication is used to set the application name. On Linux (Xorg) this can be queried using the xprop utility. Specifically, this name will be used for the ``WM_CLASS`` property [1], which can be used by X-based window managers to manipulate window placement and app-specific behaviour (automatic floating, tiling, positioning, etc). The approach proposed here is used by Jupyter's QtConsole [2] for example.

[1] https://tronche.com/gui/x/icccm/sec-4.html#WM_CLASS
[2] https://github.com/jupyter/qtconsole/pull/16