core: prune return None from inserted Python blocks
In Python <3.10+, we were sure to find a single implicit return None per python block. Starting with 3.10, the implicit return None in with or try/except can be duplicated rather than being behind a jump. So we identify all explicit return None in the AST and prune implicit return None from all blocks in the bytecode control flow graph.