widgets: reimplement initialize instead of shadowing the initialized member
Atom 0.9.0 broke this but I do not consider it a bug on atom. Shadowing an event with a method is not something I fancy supporting and the issue here looks like bad auto-completion to me since other widgets do override this method and if the state of internal private vars is relevant to initialized handlers (it should not) one can suppress the notification and re-emit initialized later.